:sparkle: Refactor CLI dir
- modelci.cli.py -> modelci.cli.__init__.py
- modelci.hub.cli.py -> modelci.cli.model_cli.py
