+++ modelci/cli/__init__.py
@@ -7,7 +7,7 @@ Date: 10/12/2020
 """
 import click
 
-import modelci.hub.cli
+import modelci.cli.model_cli
 from modelci.app import (start as app_start, stop as app_stop)
 from modelci.utils import Logger
 from modelci.utils.docker_container_manager import DockerContainerManager
@@ -50,8 +50,8 @@ def clean():
     container_conn.remove_all()
 
 
-cli.add_command(modelci.hub.cli.commands)
-cli.add_command(modelci.hub.cli.models)
+cli.add_command(modelci.cli.model_cli.commands)
+cli.add_command(modelci.cli.model_cli.models)
 
 if __name__ == '__main__':
     cli()
